requirements.txt
django
numpy
pytest
opencv-python
tensorflow
keras
torch
torchvision
matplotlib
Pillow
